+++ app/src/main/java/com/texthip/thip/ui/group/myroom/component/GroupMySectionHeader.kt
@@ -1,4 +1,4 @@
-package com.texthip.thip.ui.myPage.groupPage
+package com.texthip.thip.ui.group.myroom.component
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Row
