+++ app/src/main/java/com/texthip/thip/ui/group/myroom/mock/GroupBottomButtonType.kt
@@ -1,3 +1,3 @@
-package com.texthip.thip.ui.myPage.groupRoom
+package com.texthip.thip.ui.group.myroom.mock
 
 enum class GroupBottomButtonType { JOIN, CANCEL, CLOSE }
