+++ app/src/main/java/com/texthip/thip/ui/group/myroom/mock/GroupCardItemRoomData.kt
@@ -1,4 +1,4 @@
-package com.texthip.thip.ui.myPage.myGroup
+package com.texthip.thip.ui.group.myroom.mock
 
 import com.texthip.thip.R
