+++ app/src/main/java/com/texthip/thip/ui/group/myroom/mock/GroupDeadlineRoomData.kt
@@ -1,0 +1,6 @@
+package com.texthip.thip.ui.group.myroom.mock
+
+data class GroupDeadlineRoomData(
+    val genre: String,
+    val rooms: List<GroupCardItemRoomData>
+)
