+++ app/src/main/java/com/texthip/thip/ui/myPage/MyPageScreen.kt
@@ -1,0 +1,19 @@
+package com.texthip.thip.ui.myPage
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.navigation.NavController
+
+@Composable
+fun MyPageScreen(navController: NavController) {
+    Box(
+        modifier = Modifier.Companion.fillMaxSize(),
+        contentAlignment = Alignment.Companion.Center
+    ) {
+        Text(text = "MyPage Screen")
+    }
+}
