--- app/src/main/java/com/texthip/thip/ui/myPage/groupPage/GroupDeadlineRoomData.kt
@@ -1,8 +1,0 @@
-package com.texthip.thip.ui.myPage.groupPage
-
-import com.texthip.thip.ui.myPage.myGroup.GroupCardItemRoomData
-
-data class GroupDeadlineRoomData(
-    val genre: String,
-    val rooms: List<GroupCardItemRoomData>
-)
